requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.0
jinja2==3.1.4
